backend/requirements.txt
fastapi
uvicorn
pandas
numpy
sqlalchemy
psycopg2-binary
python-multipart
pydantic
openpyxl
pytest
hypothesis
alembic
python-json-logger
python-jose[cryptography]
passlib[bcrypt]
apscheduler
redis
prometheus_client
psutil
streamlit
plotly
requests
